refactor: move test utils to test_utils.go
